modify: Change how Eunm is used
