--- Dungeon Jumak/Assets/00. Scripts/98. Manager.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 25c5ebee7be14aa449d9c17921dc6bf1
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
