--- Dungeon Jumak/Assets/00. Scripts/98. Manager/Define.cs
@@ -1,19 +1,0 @@
-using UnityEngine;
-
-public class Define : MonoBehaviour
-{
-    public enum Sound
-    {
-        Bgm,
-        Effect,
-        MaxCount,
-    }
-
-    public enum SFX_Label
-    {
-        Main_SFX,
-        Dungeon_SFX,
-        ETC_SFX,
-    }
-
-}
