--- Dungeon Jumak/Assets/00. Scripts/98. Manager/Define.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 48914610af7e8114b94a051ec4dd832c
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
